🎨 Rearrage process interface
