+++ bluemira/codes/process/setup.py
@@ -25,6 +25,9 @@ PROCESS setup functions
 
 import os
 
+import bluemira.codes.interface as interface
+from bluemira.codes.error import CodesError
+
 # PROCESS imports
 from bluemira.codes.process.api import DEFAULT_INDAT, InDat
 
@@ -46,3 +49,71 @@ class PROCESSInputWriter(InDat):
         else:
             super().__init__(filename=None)
             self.filename = template_indat
+
+
+class Setup(interface.Setup):
+    def _run(self):
+        # Write the IN.DAT file and store in the main PROCESS folder
+        # Note that if use_bp_inputs is True, bluemira outputs with
+        # param.mapping.send == True will be written to IN.DAT.
+        self.prepare_bp_inputs()
+        self.write_indat(use_bp_inputs=True)
+
+    def _runinput(self):
+        self.prepare_bp_inputs()
+        self.write_indat(use_bp_inputs=False)
+
+    def prepare_bp_inputs(self, use_bp_inputs=True):
+        """
+        Update parameter mapping send values to True/False depending on use_bp_inputs.
+
+        Parameters
+        ----------
+        use_bp_inputs: bool, optional
+            Option to use bluemira values as PROCESS inputs. If True, sets the send
+            value for params in the params_to_update list to True and sets all others to
+            False. If True but no params_to_update list provided, makes no changes to
+            send values. If False, sets all send values to False.
+            Default, True
+        """
+        # Skip if True but no list provided
+        if use_bp_inputs is True and self.parent._params_to_update is None:
+            return
+        # Update send values to True or False
+        for param in self.parent._params.get_parameter_list():
+            bp_name = self.parent._parameter_mapping[param.mapping[PROCESS].name]
+            param.mapping[PROCESS].send = (
+                use_bp_inputs and bp_name in self.parent._params_to_update
+            )
+
+    def write_indat(self, use_bp_inputs=True):
+        """
+        Write the IN.DAT file and stores in the main PROCESS folder.
+
+        Parameters
+        ----------
+        use_bp_inputs: bool, optional
+            Option to use bluemira values as PROCESS inputs. Used to re-run PROCESS
+            within a bluemira run. If False, runs PROCESS without modifying inputs.
+            Default, True
+        """
+        # Load defaults in bluemira folder
+        writer = PROCESSInputWriter(template_indat=self.parent._template_indat)
+        if writer.data == {}:
+            raise CodesError(
+                f"Unable to read template IN.DAT file at {self.parent._template_indat}"
+            )
+
+        if use_bp_inputs is True:
+            for param in self.parent._params.get_parameter_list():
+                mapping = param.mapping[PROCESS]
+                if mapping.send:
+                    new_mapping = update_obsolete_vars(mapping.name)
+                    if isinstance(new_mapping, list):
+                        for mapping in new_mapping:
+                            writer.add_parameter(mapping, param.value)
+                    else:
+                        writer.add_parameter(new_mapping, param.value)
+
+        filename = os.path.join(self._run_dir, "IN.DAT")
+        writer.write_in_dat(output_filename=filename)
