+++ bluemira/codes/process/teardown.py
@@ -31,6 +31,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 import bluemira.base as bm_base
+import bluemira.codes.interface as interface
 from bluemira.base.look_and_feel import bluemira_warn
 from bluemira.base.parameter import ParameterFrame
 from bluemira.codes.error import CodesError
@@ -233,6 +234,142 @@ class BMFile(MFile):
                 found = True
                 break  # only keep one!
         return found
+
+
+class Teardown(interface.Teardown):
+    def _run(self):
+        self._check_PROCESS_output()
+
+        # Load PROCESS results into bluemira
+        self._load_PROCESS(self.read_mfile(), recv_all=False)
+
+    def _runinput(self):
+        self._check_PROCESS_output()
+
+        # Load PROCESS results into bluemira
+        self._load_PROCESS(self.read_mfile(), recv_all=True)
+
+    def _read(self):
+        self.get_PROCESS_run(path=self.parent._read_dir, recv_all=False)
+
+    def _readall(self):
+        self.get_PROCESS_run(path=self.parent._read_dir, recv_all=True)
+
+    def _mock(self):
+        self.mock_PROCESS_run()
+
+    def get_PROCESS_run(self, path, recv_all=False):
+        """
+        Loads an existing PROCESS file (read-only). Not to be used when running PROCESS.
+        """
+        bluemira_print("Loading PROCESS systems code run.")
+
+        # Load the PROCESS MFile & read selected output
+        params_to_recv = (
+            self.parent._parameter_mapping if recv_all else self.parent._recv_mapping
+        )
+        self._load_PROCESS(BMFile(path, params_to_recv), recv_all)
+
+    def _load_PROCESS(self, bm_file: BMFile, recv_all: bool = False):
+        """
+        Loads a PROCESS output file (MFILE.DAT) and extract some or all its output data
+
+        Parameters
+        ----------
+            bm_file: BMFile
+                PROCESS output file (MFILE.DAT) to load
+            recv_all: bool, optional
+                True - Read all PROCESS output with a mapping,
+                False - reads only PROCESS output with a mapping and recv = True.
+                Default, False
+        """
+        # Load all PROCESS vars mapped with a bluemira input
+        var = (
+            self.parent._parameter_mapping.values()
+            if recv_all
+            else self.parent._recv_mapping.values()
+        )
+        param = bm_file.extract_outputs(var)
+        self.parent._params.update_kw_parameters(dict(zip(var, param)), source=PROCESS)
+
+    def read_mfile(self):
+        """
+        Read the MFILE.DAT from the PROCESS run_dir.
+
+        Returns
+        -------
+        mfile: BMFile
+            The object representation of the output MFILE.DAT.
+        """
+        m_file = BMFile(self._run_dir, self.parent._parameter_mapping)
+        self._check_feasible_solution(m_file)
+        return m_file
+
+    def mock_PROCESS_run(self):
+        """
+        Mock PROCESS. To be used in tests and examples only!
+        """
+        bluemira_print("Mocking PROCESS systems code run")
+
+        # Create mock PROCESS file.
+        path = self.parent._read_dir
+        filename = os.sep.join([path, "mockPROCESS.json"])
+        with open(filename, "r") as fh:
+            process = json.load(fh)
+
+        self.parent._params.update_kw_parameters(process, source=f"{PROCESS} (Mock)")
+
+    def _check_PROCESS_output(self):
+        """
+        Check that PROCESS has produced valid (non-zero lined) output.
+
+        Raises
+        ------
+        CodesError
+            If any resulting output files don't exist or are empty.
+        """
+        for filename in self.parent.output_files:
+            filepath = os.sep.join([self._run_dir, filename])
+            if os.path.exists(filepath):
+                with open(filepath) as fh:
+                    if len(fh.readlines()) == 0:
+                        message = (
+                            f"PROCESS generated an empty {filename} "
+                            f"file in {self._run_dir} - check PROCESS logs."
+                        )
+                        bluemira_warn(message)
+                        raise CodesError(message)
+            else:
+                message = (
+                    f"PROCESS run did not generate the {filename} "
+                    f"file in {self._run_dir} - check PROCESS logs."
+                )
+                bluemira_warn(message)
+                raise CodesError(message)
+
+    @staticmethod
+    def _check_feasible_solution(m_file):
+        """
+        Check that PROCESS found a feasible solution.
+
+        Parameters
+        ----------
+        m_file: BMFile
+            The PROCESS MFILE to check for a feasible solution
+
+        Raises
+        ------
+        CodesError
+            If a feasible solution was not found.
+        """
+        error_code = m_file.params["Numerics"]["ifail"]
+        if error_code != 1:
+            message = (
+                f"PROCESS did not find a feasible solution. ifail = {error_code}."
+                " Check PROCESS logs."
+            )
+            bluemira_warn(message)
+            raise CodesError(message)
 
 
 def boxr(ri, ro, w, off=0):
