+++ bluemira/codes/wrapper.py
@@ -83,7 +83,7 @@ def run_systems_code(
     if (not process.PROCESS_ENABLED) and (process_mode.lower() != "mock"):
         raise CodesError("PROCESS not (properly) installed")
 
-    runner = process.Run(
+    runner = process.ProcessSolver(
         params, build_config, run_dir, read_dir, template_indat, params_to_update
     )
     return runner.params
